refact routes
