+++ apps/web/src/components/agent/DetailHeader.tsx
@@ -7,6 +7,7 @@ import { AgentAvatar } from "../common/Avatar.tsx";
 import { DockedToggleButton } from "../pageLayout.tsx";
 import { Button } from "@deco/ui/components/button.tsx";
 import { useFocusChat } from "../agents/hooks.ts";
+import { useOpenSettingsIfQueryParam } from "./chat.tsx";
 
 interface Props {
   agentId: string;
@@ -54,6 +55,7 @@ AgentHeader.Skeleton = () => {
 };
 
 AgentHeader.UI = ({ agentId }: Props) => {
+  useOpenSettingsIfQueryParam();
   const { data: agent } = useAgent(agentId);
   const focusChat = useFocusChat();
