+++ apps/web/src/components/agents/hooks.ts
@@ -2,45 +2,12 @@ import { useCallback } from "react";
 import { useNavigate } from "react-router";
 import { useBasePath } from "../../hooks/useBasePath.ts";
 
-// Helper to get agent URL
-const getAgentPath = (agentId: string, threadId?: string): string =>
-  `/agent/${agentId}/${threadId ?? ""}`;
-
 const getChatPath = (agentId: string, threadId: string): string =>
   `/chat/${agentId}/${threadId}`;
 
-interface AgentNavigationOptions {
-  threadId?: string;
-  message?: string;
-}
-
-export const useFocusAgent = () => {
-  const navigate = useNavigate();
-  const withBasePath = useBasePath();
-
-  const navigateToAgent = useCallback(
-    (agentId: string, options?: AgentNavigationOptions) => {
-      const pathname = withBasePath(getAgentPath(agentId, options?.threadId));
-
-      // Add message as a query parameter if provided
-      let url = pathname;
-      if (options?.message) {
-        const searchParams = new URLSearchParams();
-        searchParams.append("message", options.message);
-        url = `${pathname}?${searchParams.toString()}`;
-      }
-
-      // Navigate to the agent page
-      navigate(url);
-    },
-    [withBasePath, navigate],
-  );
-
-  return navigateToAgent;
-};
-
 interface ChatNavigationOptions {
   message?: string;
+  openSettings?: boolean;
 }
 
 export const useFocusChat = () => {
@@ -50,12 +17,20 @@ export const useFocusChat = () => {
   const navigateToAgent = useCallback(
     (agentId: string, threadId: string, options?: ChatNavigationOptions) => {
       const pathname = withBasePath(getChatPath(agentId, threadId));
-
-      // Add message as a query parameter if provided
+      // Add query parameters if options are provided
       let url = pathname;
+      const searchParams = new URLSearchParams();
+
       if (options?.message) {
-        const searchParams = new URLSearchParams();
         searchParams.append("message", options.message);
+      }
+
+      if (options?.openSettings) {
+        searchParams.append("openSettings", "true");
+      }
+
+      // Only append search params if we have any
+      if (searchParams.toString()) {
         url = `${pathname}?${searchParams.toString()}`;
       }
