+++ apps/web/src/main.tsx
@@ -64,10 +64,6 @@ const AgentsList = lazy(
   () => wrapWithUILoadingFallback(import("./components/agents/list.tsx")),
 );
 
-const AgentDetail = lazy(
-  () => wrapWithUILoadingFallback(import("./components/agent/detail.tsx")),
-);
-
 const Chat = lazy(
   () => wrapWithUILoadingFallback(import("./components/agent/chat.tsx")),
 );
@@ -159,10 +155,6 @@ function Router() {
         <Route
           path="agents"
           element={<AgentsList />}
-        />
-        <Route
-          path="agent/:id"
-          element={<AgentDetail />}
         />
         <Route
           path="chat/:id/:threadId"
